requirements.txt
aioconsole==0.8.1
aiofiles==24.1.0
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
altair==5.4.1
annotated-types==0.7.0
anthropic==0.42.0
anyio==4.4.0
arabic-reshaper==3.0.0
asn1crypto==1.5.1
async-timeout==4.0.3
asyncio-atexit==1.0.1
attrs==24.2.0
autogen-agentchat==0.4.0.dev11
autogen-core==0.4.0.dev11
autogen-ext==0.4.0.dev11
beautifulsoup4==4.12.3
biopython==1.84
blinker==1.8.2
cachetools==5.5.0
certifi==2024.7.4
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
cobble==0.1.4
colour==0.1.5
contourpy==1.2.1
cryptography==43.0.3
cssselect2==0.7.0
cycler==0.12.1
daal==2021.4.0
daal4py==2021.6.0
dataclasses-json==0.6.7
datasets==2.19.1
Deprecated==1.2.15
dill==0.3.8
diskcache==5.6.3
distro==1.9.0
docker==7.1.0
easyocr==1.7.2
et-xmlfile==1.1.0
exceptiongroup==1.2.2
faiss==1.8.0
filelock==3.15.4
FLAML==2.3.1
fonttools==4.53.1
frozenlist==1.5.0
fsspec==2024.3.1
geographiclib==2.0
geopy==2.4.1
git-filter-repo==2.45.0
gitdb==4.0.11
GitPython==3.1.43
google-api-core==2.19.2
google-api-python-client==2.144.0
google-auth==2.34.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
google_search_results==2.4.2
googleapis-common-protos==1.65.0
greenlet==3.1.1
h11==0.14.0
html5lib==1.1
httpcore==1.0.5
httplib2==0.22.0
httpx==0.27.0
huggingface-hub==0.24.6
idna==3.7
imageio==2.36.0
importlib_metadata==8.4.0
Jinja2==3.1.4
jiter==0.8.2
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonref==1.1.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
katex==0.0.4
kiwisolver==1.4.5
langchain==0.2.3
langchain-community==0.2.0
langchain-core==0.2.0
langchain-openai==0.1.7
langchain-text-splitters==0.2.1
langsmith==0.1.137
latex2mathml==3.77.0
lazy_loader==0.4
llvmlite==0.43.0
lxml==5.3.0
mammoth==1.7.1
markdown-it-py==3.0.0
markdownify==0.12.1
MarkupSafe==2.1.5
marshmallow==3.23.0
matplotlib==3.9.1
mdurl==0.1.2
more-itertools==10.5.0
mpmath==1.3.0
msgpack==1.1.0
multidict==6.1.0
multiprocess==0.70.16
mypy-extensions==1.0.0
narwhals==1.8.4
networkx==3.3
ninja==1.11.1.1
numba==0.60.0
numpy==1.26.4
oauthlib==3.2.2
openai==1.58.1
openai-whisper==20231117
opencv-python-headless==4.10.0.84
openpyxl==3.1.2
opentelemetry-api==1.27.0
orjson==3.10.10
oscrypto==1.3.0
packaging==23.2
pandas==2.2.2
pathvalidate==3.2.0
pdf2image==1.17.0
pdfminer==20191125
pdfminer.six==20231228
pillow==10.3.0
playwright==1.49.1
propcache==0.2.0
proto-plus==1.24.0
protobuf==4.25.5
PubChemPy==1.0.4
puremagic==1.23
pyarrow==17.0.0
pyarrow-hotfix==0.6
pyasn1==0.6.0
pyasn1_modules==0.4.0
pyautogen==0.2.27
pyclipper==1.3.0.post6
pycparser==2.22
pycryptodome==3.21.0
pydantic==2.8.2
pydantic_core==2.20.1
pydeck==0.9.1
pydub==0.25.1
pyee==12.0.0
Pygments==2.18.0
pyHanko==0.25.1
pyhanko-certvalidator==0.26.3
PyMuPDF==1.25.1
pyparsing==3.1.2
pypdf==5.1.0
PyPDF2==3.0.1
python-bidi==0.6.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-pptx==0.6.23
pytz==2024.1
PyYAML==6.0.2
qrcode==8.0
ray==2.22.0
referencing==0.35.1
regex==2024.7.24
reportlab==4.2.5
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==13.7.1
rpds-py==0.20.0
rsa==4.9
safetensors==0.4.5
scikit-image==0.24.0
scipy==1.14.1
seaborn==0.13.2
serpapi==0.1.5
shapely==2.0.6
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.6
SpeechRecognition==3.10.4
SQLAlchemy==2.0.36
streamlit==1.38.0
svglib==1.5.1
sympy==1.13.1
tbb==2021.10.0
tenacity==8.5.0
termcolor==2.5.0
tifffile==2024.9.20
tiktoken==0.8.0
tinycss2==1.4.0
tokenizers==0.20.3
toml==0.10.2
torch==2.2.2
torchaudio==2.2.2
torchvision==0.17.2
tornado==6.4.1
tqdm==4.66.5
transformers==4.46.0
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.1
tzlocal==5.2
uritemplate==4.1.1
uritools==4.0.3
urllib3==2.2.2
webencodings==0.5.1
wrapt==1.17.0
xhtml2pdf==0.2.16
XlsxWriter==3.2.0
xxhash==3.5.0
yarl==1.16.0
youtube-transcript-api==0.6.2
zipp==3.21.0
